Move src/lib/dnssd/minimal_mdns/core to common location
