+++ src/lib/dnssd/minimal_mdns/QueryBuilder.h
@@ -20,7 +20,7 @@
 #include <system/SystemPacketBuffer.h>
 
 #include <lib/dnssd/minimal_mdns/Query.h>
-#include <lib/dnssd/minimal_mdns/core/DnsHeader.h>
+#include <lib/dnssd/wire/DnsHeader.h>
 
 namespace mdns {
 namespace Minimal {
