+++ src/lib/dnssd/minimal_mdns/records/ResourceRecord.h
@@ -19,9 +19,9 @@
 
 #include <cstddef>
 
-#include <lib/dnssd/minimal_mdns/core/Constants.h>
-#include <lib/dnssd/minimal_mdns/core/QName.h>
-#include <lib/dnssd/minimal_mdns/core/RecordWriter.h>
+#include <lib/dnssd/wire/Constants.h>
+#include <lib/dnssd/wire/QName.h>
+#include <lib/dnssd/wire/RecordWriter.h>
 
 #include <lib/support/BufferWriter.h>
