+++ src/lib/dnssd/minimal_mdns/responders/Responder.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include <lib/dnssd/minimal_mdns/core/QName.h>
+#include <lib/dnssd/wire/QName.h>
 #include <lib/dnssd/minimal_mdns/records/ResourceRecord.h>
 
 #include <inet/IPPacketInfo.h>
