+++ src/lib/dnssd/wire/QName.cpp
