+++ src/lib/dnssd/wire/RecordWriter.cpp
