+++ src/lib/dnssd/wire/RecordWriter.h
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <lib/core/CHIPEncoding.h>
-#include <lib/dnssd/minimal_mdns/core/QName.h>
+#include <lib/dnssd/wire/QName.h>
 #include <lib/support/BufferWriter.h>
 
 #include <optional>
