+++ src/lib/dnssd/wire/tests/TestFlatAllocatedQName.cpp
@@ -18,7 +18,7 @@
 #include <pw_unit_test/framework.h>
 
 #include <lib/core/StringBuilderAdapters.h>
-#include <lib/dnssd/minimal_mdns/core/FlatAllocatedQName.h>
+#include <lib/dnssd/wire/FlatAllocatedQName.h>
 
 namespace {
